+++ package.js
@@ -3,19 +3,9 @@ Package.describe({
 });
 
 Package.on_use(function (api) {
-	// TODO, 0.6.4 and below support until 2014
-	try {
-	    api.use('webapp', 'server');
-	}
-	catch (error) {
-	    if (error.code != 'ENOENT')
-	        throw(error);
-	}
-
+  api.use('webapp', 'server');
 	api.use('robots-txt', 'server');
 	api.add_files('sitemaps.js', 'server');
-
-	// TODO, 0.6.4 and below support until 2014
-	if (api.export)
-		api.export('sitemaps', 'server');
+	
+	api.export('sitemaps', 'server');
 });